Fixed "cloning" bug of items through depot (#460)
Resolves #449

With the wrong size, instead of getting all possible pages, it was stopping at the eighth page, which is the maximum number of items in uint8_t, so when there were more items it allowed to clone in a specific scenario, I will avoid giving examples so that they do not abuse the bug on the server of those who are not using the fix yet

+++ src/creatures/players/player.cpp
@@ -3529,7 +3529,7 @@ void Player::removeThing(Thing* thing, uint32_t count)
 	}
 }
 
-uint8_t Player::getThingIndex(const Thing* thing) const
+int32_t Player::getThingIndex(const Thing* thing) const
 {
 	for (uint8_t i = CONST_SLOT_FIRST; i <= CONST_SLOT_LAST; ++i) {
 		if (inventory[i] == thing) {
